requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
sqlalchemy==2.0.41
anthropic==0.52.0
jinja2==3.1.6
python-dotenv==1.1.0
python-multipart==0.0.20
itsdangerous==2.2.0
ollama==0.6.1
